- B data and gps data changed to double in postgre (float8). float4 reach only 6 digits precission. - All statistics for B are implemented (including medians and standard deviations) - maby B data cheks prepared in excel file. This file needs updating.

+++ magneticter/dbsave.h
@@ -9,7 +9,7 @@
 struct gps_db_info_t
 {
     time_t utc_time;
-    float  latitute, longitude, altitude;
+    double latitute, longitude, altitude;
 
     gps_db_info_t() :
     utc_time(0), latitute(0), longitude(0), altitude(0) {}
